chore: move Type implementations to separate file
