+++ pkg/config/config.go
@@ -67,107 +67,9 @@ type Parameters map[string]parameter.Parameter
 
 type TypeID string
 
-const (
-	SettingsTypeID       TypeID = "settings"
-	ClassicApiTypeID     TypeID = "classic"
-	EntityTypeID         TypeID = "entity"
-	AutomationTypeID     TypeID = "automation"
-	BucketTypeID         TypeID = "bucket"
-	DocumentTypeID       TypeID = "document"
-	OpenPipelineTypeID   TypeID = "openpipeline"
-	GrailFilterSegmentID TypeID = "filter-segment"
-)
-
 type Type interface {
 	// ID returns the type-id.
 	ID() TypeID
-}
-
-type SettingsType struct {
-	SchemaId, SchemaVersion string
-}
-
-func (SettingsType) ID() TypeID {
-	return SettingsTypeID
-}
-
-type ClassicApiType struct {
-	Api string
-}
-
-func (ClassicApiType) ID() TypeID {
-	return ClassicApiTypeID
-}
-
-type EntityType struct {
-	EntitiesType string
-}
-
-func (EntityType) ID() TypeID {
-	return EntityTypeID
-}
-
-// AutomationResource defines which resource is an AutomationType
-type AutomationResource string
-
-const (
-	Workflow         AutomationResource = "workflow"
-	BusinessCalendar AutomationResource = "business-calendar"
-	SchedulingRule   AutomationResource = "scheduling-rule"
-)
-
-// AutomationType represents any Dynatrace Platform automation-resource
-type AutomationType struct {
-	// Resource identifies which Automation resource is used in this config.
-	// Currently, this can be Workflow, BusinessCalendar, or SchedulingRule.
-	Resource AutomationResource
-}
-
-func (AutomationType) ID() TypeID {
-	return AutomationTypeID
-}
-
-type BucketType struct{}
-
-func (BucketType) ID() TypeID {
-	return BucketTypeID
-}
-
-// DocumentKind defines the type of a document. Currently, it can be a dashboard or a notebook.
-type DocumentKind string
-
-const (
-	DashboardKind DocumentKind = "dashboard"
-	NotebookKind  DocumentKind = "notebook"
-)
-
-// DocumentType represents a Dynatrace platform document.
-type DocumentType struct {
-	// Kind indicates the type of document.
-	Kind DocumentKind
-
-	// Private indicates if a document is private, otherwise by default it is visible to other users.
-	Private bool
-}
-
-func (DocumentType) ID() TypeID {
-	return DocumentTypeID
-}
-
-// OpenPipelineType represents an OpenPipeline configuration.
-type OpenPipelineType struct {
-	// Kind indicates the type of OpenPipeline.
-	Kind string
-}
-
-func (OpenPipelineType) ID() TypeID {
-	return OpenPipelineTypeID
-}
-
-type GrailFilterSegment struct{}
-
-func (GrailFilterSegment) ID() TypeID {
-	return GrailFilterSegmentID
 }
 
 // Config struct defining a configuration which can be deployed.
